# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Naytia 像素小站 API",
  "description": "账号 / 论坛 / 工具箱收藏。契约见 docs/plans/03。",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/reverse/overview": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "Get Overview",
    "operationId": "get_overview_api_reverse_overview_get",
    "parameters": [
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "IDA 实例端口；省略则用当前实例",
       "title": "Port"
      },
      "description": "IDA 实例端口；省略则用当前实例"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BinaryOverview"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/functions": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "List Functions",
    "operationId": "list_functions_api_reverse_functions_get",
    "parameters": [
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "count",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 100,
       "title": "Count"
      }
     },
     {
      "name": "filter",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 200
        },
        {
         "type": "null"
        }
       ],
       "description": "名称 glob，如 *Player*",
       "title": "Filter"
      },
      "description": "名称 glob，如 *Player*"
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FunctionPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/functions/lookup": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "Lookup Functions",
    "operationId": "lookup_functions_api_reverse_functions_lookup_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "minItems": 1,
       "maxItems": 50,
       "description": "地址或函数名，可重复",
       "title": "Q"
      },
      "description": "地址或函数名，可重复"
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/FunctionSummary"
         },
         "title": "Response Lookup Functions Api Reverse Functions Lookup Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/decompile": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "Decompile",
    "operationId": "decompile_api_reverse_decompile_get",
    "parameters": [
     {
      "name": "addr",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 200,
       "description": "函数地址或名称",
       "title": "Addr"
      },
      "description": "函数地址或名称"
     },
     {
      "name": "include_addresses",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Include Addresses"
      }
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Pseudocode"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/disasm": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "Disasm",
    "operationId": "disasm_api_reverse_disasm_get",
    "parameters": [
     {
      "name": "addr",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 200,
       "title": "Addr"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "max_instructions",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 50000,
       "minimum": 1,
       "default": 2000,
       "title": "Max Instructions"
      }
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Disassembly"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/xrefs": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "Xrefs",
    "operationId": "xrefs_api_reverse_xrefs_get",
    "parameters": [
     {
      "name": "addr",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 200,
       "title": "Addr"
      }
     },
     {
      "name": "direction",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(to|from|both)$",
       "default": "to",
       "title": "Direction"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/XrefGroup"
         },
         "title": "Response Xrefs Api Reverse Xrefs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/callees": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "Callees",
    "operationId": "callees_api_reverse_callees_get",
    "parameters": [
     {
      "name": "addr",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 200,
       "title": "Addr"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 200,
       "title": "Limit"
      }
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/CalleeGroup"
         },
         "title": "Response Callees Api Reverse Callees Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/blocks": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "Blocks",
    "operationId": "blocks_api_reverse_blocks_get",
    "parameters": [
     {
      "name": "addr",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 200,
       "title": "Addr"
      }
     },
     {
      "name": "max_blocks",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 10000,
       "minimum": 1,
       "default": 500,
       "title": "Max Blocks"
      }
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/BlockGroup"
         },
         "title": "Response Blocks Api Reverse Blocks Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/search/strings": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "Search Strings",
    "operationId": "search_strings_api_reverse_search_strings_get",
    "parameters": [
     {
      "name": "pattern",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 200,
       "title": "Pattern"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StringSearchResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/search/text": {
   "get": {
    "tags": [
     "reverse"
    ],
    "summary": "Search Text",
    "operationId": "search_text_api_reverse_search_text_get",
    "parameters": [
     {
      "name": "pattern",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 2,
       "maxLength": 200,
       "title": "Pattern"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 200
        },
        {
         "type": "null"
        }
       ],
       "title": "Start"
      }
     },
     {
      "name": "regex",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Regex"
      }
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TextSearchResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/reverse/functions/rename": {
   "post": {
    "tags": [
     "reverse"
    ],
    "summary": "Rename Function",
    "description": "重命名函数。权限在服务端判定（01 文档第 7 节）；前端隐藏按钮不算权限控制。",
    "operationId": "rename_function_api_reverse_functions_rename_post",
    "parameters": [
     {
      "name": "addr",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 200,
       "title": "Addr"
      }
     },
     {
      "name": "port",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Port"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RenameRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response Rename Function Api Reverse Functions Rename Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BinaryOverview": {
    "properties": {
     "module": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Module"
     },
     "path": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Path"
     },
     "arch": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Arch"
     },
     "base_address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Base Address"
     },
     "image_size": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Image Size"
     },
     "total_functions": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total Functions"
     },
     "named_functions": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Named Functions"
     },
     "library_functions": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Library Functions"
     },
     "unnamed_functions": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Unnamed Functions"
     },
     "total_strings": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total Strings"
     },
     "total_segments": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total Segments"
     },
     "segments": {
      "items": {
       "$ref": "#/components/schemas/SegmentInfo"
      },
      "type": "array",
      "title": "Segments"
     }
    },
    "type": "object",
    "title": "BinaryOverview",
    "description": "survey_binary 的归一化结果。"
   },
...
   "BlockGroup": {
    "properties": {
     "addr": {
      "type": "string",
      "title": "Addr"
     },
     "blocks": {
      "items": {
       "$ref": "#/components/schemas/BasicBlock"
      },
      "type": "array",
      "title": "Blocks"
     },
     "count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Count"
     },
     "total_blocks": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total Blocks"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     }
    },
    "type": "object",
    "required": [
     "addr"
    ],
    "title": "BlockGroup"
   },
...
   "CalleeGroup": {
    "properties": {
     "addr": {
      "type": "string",
      "title": "Addr"
     },
     "callees": {
      "items": {
       "$ref": "#/components/schemas/CalleeItem"
      },
      "type": "array",
      "title": "Callees"
     },
     "more": {
      "type": "boolean",
      "title": "More",
      "default": false
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     }
    },
    "type": "object",
    "required": [
     "addr"
    ],
    "title": "CalleeGroup"
   },
   "CalleeItem": {
    "properties": {
     "addr": {
      "type": "string",
      "title": "Addr"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type"
     }
    },
    "type": "object",
    "required": [
     "addr"
    ],
    "title": "CalleeItem"
   },
...
   "DisasmLine": {
    "properties": {
     "addr": {
      "type": "string",
      "title": "Addr"
     },
     "instruction": {
      "type": "string",
      "title": "Instruction"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     }
    },
    "type": "object",
    "required": [
     "addr",
     "instruction"
    ],
    "title": "DisasmLine"
   },
   "Disassembly": {
    "properties": {
     "addr": {
      "type": "string",
      "title": "Addr"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "start_ea": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Ea"
     },
     "segment": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Segment"
     },
     "lines": {
      "items": {
       "$ref": "#/components/schemas/DisasmLine"
      },
      "type": "array",
      "title": "Lines"
     },
     "stack_frame": {
      "items": {
       "$ref": "#/components/schemas/StackVar"
      },
      "type": "array",
      "title": "Stack Frame"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     }
    },
    "type": "object",
    "required": [
     "addr"
    ],
    "title": "Disassembly"
   },
...
   "FunctionPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/FunctionSummary"
      },
      "type": "array",
      "title": "Items"
     },
     "next_offset": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Offset"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     }
    },
    "type": "object",
    "title": "FunctionPage"
   },
   "FunctionSummary": {
    "properties": {
     "addr": {
      "type": "string",
      "title": "Addr"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "size": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Size"
     }
    },
    "type": "object",
    "required": [
     "addr",
     "name"
    ],
    "title": "FunctionSummary"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Pseudocode": {
    "properties": {
     "addr": {
      "type": "string",
      "title": "Addr"
     },
     "code": {
      "type": "string",
      "title": "Code",
      "default": ""
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     }
    },
    "type": "object",
    "required": [
     "addr"
    ],
    "title": "Pseudocode"
   },
...
   "RenameRequest": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     },
     "dry_run": {
      "type": "boolean",
      "title": "Dry Run",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "RenameRequest"
   },
...
   "SegmentInfo": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "start": {
      "type": "string",
      "title": "Start"
     },
     "end": {
      "type": "string",
      "title": "End"
     },
     "size": {
      "type": "string",
      "title": "Size"
     },
     "permissions": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Permissions"
     }
    },
    "type": "object",
    "required": [
     "name",
     "start",
     "end",
     "size"
    ],
    "title": "SegmentInfo"
   },
...
   "StackVar": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "offset": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Offset"
     },
     "size": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Size"
     },
     "type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "StackVar"
   },
...
   "StringSearchResult": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total",
      "default": 0
     },
     "matches": {
      "items": {
       "$ref": "#/components/schemas/StringMatch"
      },
      "type": "array",
      "title": "Matches"
     }
    },
    "type": "object",
    "title": "StringSearchResult"
   },
...
   "TextSearchResult": {
    "properties": {
     "hits": {
      "items": {
       "$ref": "#/components/schemas/TextSearchHit"
      },
      "type": "array",
      "title": "Hits"
     },
     "next_start": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Start"
     }
    },
    "type": "object",
    "title": "TextSearchResult"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "XrefGroup": {
    "properties": {
     "addr": {
      "type": "string",
      "title": "Addr"
     },
     "xrefs": {
      "items": {
       "$ref": "#/components/schemas/XrefItem"
      },
      "type": "array",
      "title": "Xrefs"
     },
     "more": {
      "type": "boolean",
      "title": "More",
      "default": false
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     }
    },
    "type": "object",
    "required": [
     "addr"
    ],
    "title": "XrefGroup"
   },
   "XrefItem": {
    "properties": {
     "addr": {
      "type": "string",
      "title": "Addr"
     },
     "type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type"
     },
     "fn": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fn"
     },
     "fn_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fn Name"
     }
    },
    "type": "object",
    "required": [
     "addr"
    ],
    "title": "XrefItem"
   }
  }
 }
}